moving include/freerdp/debug.h to include/freerdp/utils/debug.h
