+++ include/freerdp/utils/debug.h
