+++ libfreerdp-core/cache.h
@@ -21,7 +21,7 @@
 #define __CACHE_H
 
 #include "types.h"
-#include <freerdp/debug.h>
+#include <freerdp/utils/debug.h>
 #include <freerdp/utils/memory.h>
 #include <freerdp/utils/datablob.h>
