+++ libfreerdp-core/secure.c
@@ -25,8 +25,8 @@
 #include "iso.h"
 #include "tcp.h"
 #include "types.h"
-#include <freerdp/debug.h>
 #include <freerdp/rdpset.h>
+#include <freerdp/utils/debug.h>
 #include <freerdp/utils/memory.h>
 
 #ifndef DISABLE_TLS
