+++ libfreerdp-core/stream.h
@@ -22,7 +22,7 @@
 
 #include <stddef.h>
 #include "frdp.h"
-#include <freerdp/debug.h>
+#include <freerdp/utils/debug.h>
 
 #if !(defined(L_ENDIAN) || defined(B_ENDIAN))
 #warning no endian defined
